Updated code and code structure
Changed the overall structure of the project and cleaned up the code to make it more redable and be more reusable.

the code generator function now has an integer return status, which allows it to be used both as a tool and as a function from within other programs.

the strings in the code generator were given more significant names, making it easier to understand what is going on.

Divided the examples in multiple files and added showcase files for 30, 100 and 1000 FOR_EACH macro iterations.