Added support for enumeration
Added a new macro: FOR_EACH_ENUMERATE

In older examples, the FOR_EACH_GET macro was used to get the elements of the __VA_ARGS__ list to generate a string enum automatically. This is possible because in C, list initializers used for arrays, as well as enums, allow the last element to have a comma and no further elements un the array initializer list or enum.

This is not the case for raw C functions that take variable arguments, as those strictly require the last argument to have no further commas.

The new macro FOR_EACH_ENUMERATE allows for a behavior similar to FOR_EACH_GET, tho it automatically appends a comma at the end of every element looped in the __VA_ARGS__ except the last one.